Separate AI provider registry modules
