+++ ddev/src/ddev/ai/agent/build.py
@@ -5,25 +5,16 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from functools import cached_property
 from typing import TYPE_CHECKING, Any, Protocol
 
-import anthropic
-
-from ddev.ai.agent.anthropic_client import AnthropicAgent
 from ddev.ai.agent.base import BaseAgent
+from ddev.ai.agent.registry import AgentProviderRegistry
 from ddev.ai.config.models import AgentConfig
 from ddev.ai.tools.fs.file_registry import FileRegistry
 from ddev.ai.tools.registry import ToolRegistry
 
 if TYPE_CHECKING:
     from ddev.ai.react.factory import ReActProcessFactory
-
-
-class AgentProviderConfig(Protocol):
-    """App configuration needed to create the provider registry."""
-
-    anthropic_api_key: str | None
 
 
 @dataclass(frozen=True)
@@ -43,105 +34,6 @@ class AgentRuntimeFactoryProtocol(Protocol):
         owner_id: str,
         process_factory: ReActProcessFactory,
     ) -> AgentRuntime: ...
-
-
-class AgentProvider(Protocol):
-    """Provider-specific agent construction and configuration validation."""
-
-    def validate_config(self, agent_config: AgentConfig) -> None: ...
-
-    def build_agent(
-        self,
-        agent_config: AgentConfig,
-        *,
-        tools: ToolRegistry,
-        system_prompt: str,
-        owner_id: str,
-    ) -> BaseAgent[Any]: ...
-
-
-class AgentProviderRegistry:
-    """Dispatches agent construction to configured providers."""
-
-    def __init__(self) -> None:
-        self._providers: dict[str, AgentProvider] = {}
-
-    def register(self, name: str, provider: AgentProvider) -> None:
-        if name in self._providers:
-            raise ValueError(f"Agent provider {name!r} is already registered")
-        self._providers[name] = provider
-
-    def contains(self, name: str) -> bool:
-        return name in self._providers
-
-    def _get_provider(self, name: str) -> AgentProvider:
-        provider = self._providers.get(name)
-        if provider is None:
-            raise ValueError(f"Agent provider {name!r} is not available")
-        return provider
-
-    def validate_config(self, agent_config: AgentConfig) -> None:
-        self._get_provider(agent_config.provider).validate_config(agent_config)
-
-    def build_agent(
-        self,
-        agent_config: AgentConfig,
-        *,
-        tools: ToolRegistry,
-        system_prompt: str,
-        owner_id: str,
-    ) -> BaseAgent[Any]:
-        provider = self._get_provider(agent_config.provider)
-        provider.validate_config(agent_config)
-        return provider.build_agent(
-            agent_config,
-            tools=tools,
-            system_prompt=system_prompt,
-            owner_id=owner_id,
-        )
-
-
-class AnthropicProvider:
-    """Builds Anthropic agents and lazily owns their shared SDK client."""
-
-    def __init__(self, api_key: str) -> None:
-        self._api_key = api_key
-
-    @cached_property
-    def client(self) -> anthropic.AsyncAnthropic:
-        return anthropic.AsyncAnthropic(api_key=self._api_key)
-
-    def validate_config(self, agent_config: AgentConfig) -> None:
-        """Validate Anthropic-specific agent configuration."""
-
-    def build_agent(
-        self,
-        agent_config: AgentConfig,
-        *,
-        tools: ToolRegistry,
-        system_prompt: str,
-        owner_id: str,
-    ) -> BaseAgent[Any]:
-        kwargs: dict[str, Any] = {}
-        if agent_config.model is not None:
-            kwargs["model"] = agent_config.model
-        if agent_config.max_tokens is not None:
-            kwargs["max_tokens"] = agent_config.max_tokens
-        return AnthropicAgent(
-            client=self.client,
-            tools=tools,
-            system_prompt=system_prompt,
-            name=owner_id,
-            **kwargs,
-        )
-
-
-def build_agent_provider_registry(config: AgentProviderConfig) -> AgentProviderRegistry:
-    """Build the provider registry from ddev app configuration."""
-    registry = AgentProviderRegistry()
-    if config.anthropic_api_key:
-        registry.register("anthropic", AnthropicProvider(config.anthropic_api_key))
-    return registry
 
 
 class AgentRuntimeBuilder(Protocol):
